+++ sdk/examples/wallet/06_send_micro_transaction.rs
@@ -7,7 +7,6 @@
 //! `cargo run --example send_micro_transaction --release`
 
 use iota_sdk::{
-    types::block::address::Bech32Address,
     wallet::{account::TransactionOptions, Result, SendAmountParams},
     Wallet,
 };
@@ -32,9 +31,9 @@ async fn main() -> Result<()> {
 
     // Send a micro transaction with amount 1
     let outputs = vec![SendAmountParams::new(
-        Bech32Address::try_from_str("rms1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluaw60xu")?,
+        "rms1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluaw60xu",
         1,
-    )];
+    )?];
 
     let transaction = account
         .send_amount(