+++ sdk/examples/wallet/08_send_nft.rs
@@ -7,7 +7,6 @@
 //! `cargo run --example send_nft --release`
 
 use iota_sdk::{
-    types::block::address::Bech32Address,
     wallet::{Result, SendNftParams},
     Wallet,
 };
@@ -32,10 +31,10 @@ async fn main() -> Result<()> {
             .set_stronghold_password(&std::env::var("STRONGHOLD_PASSWORD").unwrap())
             .await?;
 
-        let outputs = vec![SendNftParams {
-            address: Bech32Address::try_from_str("rms1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluaw60xu")?,
-            nft_id: *nft_id,
-        }];
+        let outputs = vec![SendNftParams::new(
+            "rms1qpszqzadsym6wpppd6z037dvlejmjuke7s24hm95s9fg9vpua7vluaw60xu",
+            *nft_id,
+        )?];
 
         let transaction = account.send_nft(outputs, None).await?;
         println!("Transaction sent: {}", transaction.transaction_id);