+++ sdk/src/wallet/wallet/builder.rs
@@ -192,15 +192,23 @@ where
         }
 
         if self.coin_type.is_none() {
-            let coin_type =
-                read_manager_builder
-                    .and_then(|data| data.coin_type)
-                    .ok_or(crate::wallet::Error::MissingParameter(
-                        "coin_type (IOTA: 4218, Shimmer: 4219)",
-                    ))?;
+            self.coin_type = read_manager_builder.and_then(|builder| builder.coin_type);
+        }
+        let coin_type = self.coin_type.ok_or(crate::wallet::Error::MissingParameter(
+            "coin_type (IOTA: 4218, Shimmer: 4219)",
+        ))?;
 
-            // Update self so it gets used and stored again
-            self.coin_type.replace(coin_type);
+        #[cfg(feature = "storage")]
+        let mut accounts = storage_manager.get_accounts().await?;
+        // Check against potential account coin type before saving the wallet data
+        #[cfg(feature = "storage")]
+        if let Some(account) = accounts.first() {
+            if *account.coin_type() != coin_type {
+                return Err(crate::wallet::Error::InvalidCoinType {
+                    new_coin_type: coin_type,
+                    existing_coin_type: *account.coin_type(),
+                });
+            }
         }
 
         // Store wallet data in storage
@@ -209,9 +217,6 @@ where
 
         #[cfg(feature = "events")]
         let event_emitter = tokio::sync::RwLock::new(EventEmitter::new());
-
-        #[cfg(feature = "storage")]
-        let mut accounts = storage_manager.get_accounts().await?;
 
         // It happened that inputs got locked, the transaction failed, but they weren't unlocked again, so we do this
         // here
@@ -227,9 +232,7 @@ where
                 .ok_or(crate::wallet::Error::MissingParameter("client_options"))?
                 .finish()
                 .await?,
-            coin_type: AtomicU32::new(self.coin_type.ok_or(crate::wallet::Error::MissingParameter(
-                "coin_type (IOTA: 4218, Shimmer: 4219)",
-            ))?),
+            coin_type: AtomicU32::new(coin_type),
             secret_manager: self
                 .secret_manager
                 .ok_or(crate::wallet::Error::MissingParameter("secret_manager"))?,